requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.5.2
frozenlist==1.7.0
gazpacho==1.1
idna==3.6
lxml==5.3.1
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
requests==2.31.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.20.1
